Instead of allowing and parsing a single-element list for UDF output_type, raise error for all backends
