+++ ibis/udf/validate.py
@@ -1,0 +1,71 @@
+"""Validation for UDFs.
+
+Warning: This is an experimental module and API here can change without notice.
+
+DO NOT USE DIRECTLY.
+"""
+
+from inspect import Parameter, signature
+
+import ibis.common.exceptions as com
+
+
+def _parameter_count(funcsig):
+    """Get the number of positional-or-keyword or position-only parameters in a
+    function signature.
+
+    Parameters
+    ----------
+    funcsig : inspect.Signature
+        A UDF signature
+
+    Returns
+    -------
+    int
+        The number of parameters
+    """
+    return sum(
+        param.kind in {param.POSITIONAL_OR_KEYWORD, param.POSITIONAL_ONLY}
+        for param in funcsig.parameters.values()
+        if param.default is Parameter.empty
+    )
+
+
+def validate_input_type_count(input_type, func):
+    """Check that the declared number of inputs (the length of `input_type`)
+    and the number of inputs to `func` are equal.
+
+    Parameters
+    ----------
+    input_type : List[DataType]
+    func : callable
+
+    Returns
+    -------
+    inspect.Signature
+    """
+    funcsig = signature(func)
+    declared_parameter_count = len(input_type)
+    function_parameter_count = _parameter_count(funcsig)
+
+    if declared_parameter_count != function_parameter_count:
+        raise TypeError(
+            'Function signature {!r} has {:d} parameters, '
+            'input_type has {:d}. These must match. Non-column '
+            'parameters must be defined as keyword only, i.e., '
+            'def foo(col, *, function_param).'.format(
+                func.__name__,
+                function_parameter_count,
+                declared_parameter_count,
+            )
+        )
+    return funcsig
+
+
+def validate_output_type(output_type):
+    """Check that the output type is a single datatype."""
+
+    if isinstance(output_type, list):
+        raise com.IbisTypeError(
+            'The output type of a UDF must be a single datatype.'
+        )
